+++ workers/unity/Packages/com.improbable.gdk.core/Worker/OpListConverter.cs
@@ -32,6 +32,7 @@ namespace Improbable.Gdk.Core
                         break;
                     case OpType.FlagUpdate:
                         var flagOp = opList.GetFlagUpdateOp(i);
+                        viewDiff.SetWorkerFlag(flagOp.Name, flagOp.Value);
                         break;
                     case OpType.LogMessage:
                         var logOp = opList.GetLogMessageOp(i);