+++ CHANGELOG.md
@@ -18,6 +18,10 @@
 - Added a `Improbable.Gdk.Core.Collections.Result<T, E>` struct to represent a result which can either contain a value `T` or an error `E`.
 - Added Scripting Define Symbol `DISABLE_REACTIVE_COMPONENTS`. Using this symbol will disable all reactive componts and systems.
     - Currently not compatible with the FPS Starter Project.
+- Added a `WorkerFlagReader` which you can subscribe and `Require`. This allows you to:
+    - Add callbacks for changes to worker flags.
+    - Read the value of worker flags.
+- Expose `GetWorkerFlag(string name)` on the `View`.
 
 ### Changed
 
@@ -35,6 +39,7 @@
 - Tools package now uses PackageManager API instead of parsing manifest.json.
 - Updated default snapshot to have more than one PlayerCreator entity.
 - Fixed package dependencies.
+- Worker flag changes are propagated to the `ViewDiff`.
 
 ## `0.2.0` - 2019-03-18